docs/Sources/google trends/backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
pytrends==4.9.2
pandas==2.1.1
numpy==1.26.0
python-dotenv==1.0.0
celery==5.3.4
SQLAlchemy==2.0.21
requests==2.31.0
gunicorn==21.2.0
marshmallow==3.20.1 